requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
google-generativeai>=0.3.0
pymupdf>=1.23.0
pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
